pyproject.toml
[tool.poetry]
name = "DSMR-reader"
version = "5"
description = "DSMR-protocol reader, telegram data storage and energy consumption visualizer"
license = "Creative Commons Attribution-NonCommercial 4.0 International Public License"
readme = 'README.md'
repository = "https://github.com/dsmrreader/dsmr-reader"
homepage = "https://dsmr-reader.readthedocs.io/"
authors = [
    "Dennis Siemensma",
    "https://github.com/dsmrreader/dsmr-reader/graphs/contributors"
]


[tool.poetry.dependencies]
python = "^3.7 || ^3.8 || ^3.9 || ^3.10"

autopep8 = "^1.5"
coreapi = "^2.3"
crcmod = "^1.7"
django = "^3.2"
django-admin-rangefilter = "^0.8"
django-admin-sortable = "^2.2"
django-colorfield = "^0.4"
django-debug-toolbar = "^3.2"
django-filter = "^21.1"
django-mathfilters = "^1.0"
djangorestframework = "^3.12"
django-solo = "^1.2"
dropbox = "^11.21"
gunicorn = "^20.1"
influxdb-client = "^1.23"
paho-mqtt = "^1.5"
psycopg2-binary = "^2.8"
pyserial = "^3.5"
pyserial-asyncio = "^0.6"
python-dateutil = "^2.8"
python-decouple = "^3.5"
pytz = "^2021.3"
pyyaml = "^6.0"
requests = "^2.26"
urllib3 = "^1.26"


[tool.poetry.dev-dependencies]
polib = "^1.1"
pylama = "^7.7"
pytest-cov = "^3.0"
pytest-django = "^4.4"
pytest-xdist = "^2.4"
mysqlclient = "^2.0"
sphinx = "^4.2"
sphinx-autobuild = "^2021.3"
sphinx-intl = "^2.0"
sphinx-rtd-theme = "^1.0"
